+++ Configurations/00-base-templates.conf
@@ -279,17 +279,17 @@
     ppc32_asm => {
 	template	=> 1,
 	cpuid_asm_src   => "ppccpuid.s ppccap.c",
-	bn_asm_src      => "bn-ppc.s ppc-mont.s",
+	bn_asm_src      => "bn-ppc.s ppc-mont.s bn_ppccap.c",
 	aes_asm_src     => "aes_core.c aes_cbc.c aes-ppc.s vpaes-ppc.s aesp8-ppc.s",
-	sha1_asm_src    => "sha1-ppc.s sha256-ppc.s sha512-ppc.s sha256p8-ppc.s sha512p8-ppc.s",
+	sha1_asm_src    => "sha1-ppc.s sha256-ppc.s sha512-ppc.s sha256p8-ppc.s sha512p8-ppc.s sha_ppccap.c",
 	modes_asm_src   => "ghashp8-ppc.s",
-	chacha_asm_src	=> "chacha-ppc.s",
-	poly1305_asm_src=> "poly1305-ppc.s poly1305-ppcfp.s",
+	chacha_asm_src	=> "chacha-ppc.s chacha_ppccap.c",
+	poly1305_asm_src=> "poly1305-ppc.s poly1305-ppcfp.s poly1305_ppccap.c",
     },
     ppc64_asm => {
 	inherit_from	=> [ "ppc32_asm" ],
 	template	=> 1,
-	ec_asm_src	=> "ecp_nistz256.c ecp_nistz256-ppc64.s",
+	ec_asm_src	=> "ecp_nistz256.c ecp_nistz256-ppc64.s ec_ppccap.c",
 
     },
 );
